[1603] Move Agent to uapi application
