--- vitrina/orgs/__init__.py
@@ -1,7 +1,0 @@
-from django.db.models import TextChoices
-from django.utils.translation import gettext_lazy as _
-
-
-class AgentType(TextChoices):
-    SPINTA = "SPINTA", "Spinta"
-    OTHER = "OTHER", _("Kita")
