+++ vitrina/orgs/templates/vitrina/orgs/tabs.html
@@ -32,8 +32,8 @@
                 </a>
             </li>
         {% endif %}
-        <li class="{% if request.resolver_match.url_name == 'organization-agents-list' %} is-active {% endif %}">
-            <a href="{% url 'organization-agents-list' organization_id %}">
+        <li class="{% if request.resolver_match.url_name == 'agent-list' %} is-active {% endif %}">
+            <a href="{% url 'agent-list' organization_id %}">
                 {% translate "Agentai" %}
             </a>
         </li>
